Interface changes refactoring and revert move of setup function
